setup friends+likes

+++ controllers/posts_controller.js
@@ -1,5 +1,7 @@
 const Post = require('../models/posts');
 const Comment = require('../models/comments');
+const Like = require('../models/likes');
+const { likes } = require('./users_controllers');
 
 module.exports.create_post = async function(req, res){
     try{
@@ -8,11 +10,14 @@ module.exports.create_post = async function(req, res){
             user: req.user._id
         });
 
-        // check if AJAX req
+        // check if AJAX req 
         if(req.xhr){
+
+            post = await post.populate('user', 'name').execPopulate();
+
             return res.status(200).json({
                 data: {
-                    post: post
+                    post: post 
                 },
                 message: "Post created!"
             });
@@ -25,7 +30,7 @@ module.exports.create_post = async function(req, res){
 
     }catch(err){
         req.flash('error', 'Cannot Post!');
-        return;
+        return res.redirect('back');
     }
 };
 
@@ -34,26 +39,35 @@ module.exports.delete_post = async function(req, res){
         let post = await Post.findById(req.params.id);
 // .id means converting id in string format
 // also the first security check that user deleting the post is the owner of the post      
-        // if(post.user == req.user.id){
-        post.remove();
+        if(post.user == req.user.id){
 
-        await Comment.deleteMany({post: req.params.id});
+            await Like.deleteMany({likeable: post, onModel: 'Post'});
+            await Like.deleteMany({_id: {$in: post.comments}});
 
-        if(req.xhr){
-            return res.status(200).json({
-                data: {
-                    post_id: req.params.id
-                },
-                message: 'Post Deleted!'
-            });
+
+            post.remove();
+
+            await Comment.deleteMany({post: req.params.id});
+
+            if(req.xhr){
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: 'Post Deleted!'
+                });
+            }
+
+
+            req.flash('success', 'Post Deleted!');
+            return res.redirect('back');
+        }else{
+            req.flash('error', 'You cannot delete this post!');
+            return res.redirect('back');
         }
-
-
-        req.flash('success', 'Post Deleted!');
-        return res.redirect('back');
 
     }catch(err){
         req.flash('error', 'Cannot Delete!');
-        return;
+        return res.redirect('back');
     };
 };
